+++ robotcode/language_server/common/server.py
@@ -1,5 +1,5 @@
 import abc
-from typing import TypeVar
+from typing import Optional, TypeVar
 
 from ...jsonrpc2.server import JsonRPCServer, JsonRpcServerMode, TcpParams
 from ...utils.logging import LoggingDescriptor
@@ -19,11 +19,9 @@ class LanguageServerBase(JsonRPCServer[TProtocol], abc.ABC):
         self,
         mode: JsonRpcServerMode = JsonRpcServerMode.STDIO,
         tcp_params: TcpParams = TcpParams(None, TCP_DEFAULT_PORT),
+        pipe_name: Optional[str] = None,
     ):
-        super().__init__(
-            mode=mode,
-            tcp_params=tcp_params,
-        )
+        super().__init__(mode=mode, tcp_params=tcp_params, pipe_name=pipe_name)
 
     @abc.abstractmethod
     def create_protocol(self) -> TProtocol:  # pragma: no cover