+++ vscode-client/utils.ts
@@ -4,24 +4,6 @@ export async function sleep(timeout: number): Promise<number> {
   });
 }
 
-export async function waitForPromise<T>(promise: Promise<T>, timeout: number): Promise<T | null> {
-  // Set a timer that will resolve with null
-  return new Promise<T | null>((resolve, reject) => {
-    const timer = setTimeout(() => resolve(null), timeout);
-    promise
-      .then((result) => {
-        // When the promise resolves, make sure to clear the timer or
-        // the timer may stick around causing tests to wait
-        clearTimeout(timer);
-        resolve(result);
-      })
-      .catch((e) => {
-        clearTimeout(timer);
-        reject(e);
-      });
-  });
-}
-
 export class Mutex {
   private mutex = Promise.resolve();